sip/zmq_logging_subscriber/requirements.txt
PyYAML>=3.12
pyzmq>=16.0.2
requests>=2.18.4
bottle>=0.12.13
bjoern>=2.0.1
